requirements.txt
requests>=2.31.0
feedparser>=6.0.10
newspaper3k>=0.2.8
google-generativeai>=0.3.0
firebase-admin>=6.2.0
beautifulsoup4>=4.12.2
exa-py>=0.6.0
googlenewsdecoder==0.1.6
python-dotenv>=1.0.0 